// src/lib/theme.d.ts
export type Theme = {
  background: string;
  surface: string;
  text: string;
  muted: string;
  primary: string;
  primaryContrast: string;
  accent: string;
  accentContrast: string;
  inverse: string;
  inverseText: string;
  inverseMuted: string;
  border: string;
  error: string;
  warning: string;
  success: string;
  cool: string;
};

export const theme: Theme;